chore: tighten the CLI surface, the secrets, and the sweep
Follow-ups from the review that are small on their own.

- refuse --config and --env on `init` instead of parsing and dropping them
- give daemon and health the storage hint their siblings already had
- report a kept local copy as info, not as a warning
- lock a workspace before sweeping, and spare directories younger than a
  minute, closing the window where a concurrent run could delete a live one
- redact the secret in credential Debug output, which Settings reaches
- return u8 from ExitStatus::code so the entrypoint stops casting
- assert the failure class in three tests that only checked for a non-empty
  message, and test the --keep mapping the dispatcher actually calls
- add Cli::command().debug_assert() and parse tests for the grammar
- narrow seven helpers to pub(super)
- add the missing LICENSE file and a CI workflow
- drop the duplicate tempfile dev-dependency

+++ src/main.rs
@@ -17,14 +17,14 @@ async fn main() -> ExitCode {
             for line in report.lines {
                 println!("{line}");
             }
-            ExitCode::from(report.status.code() as u8)
+            ExitCode::from(report.status.code())
         }
         Err(cli_err) => {
             eprintln!("error: {cli_err}");
             if let Some(hint) = cli_err.hint() {
                 eprintln!("hint:  {hint}");
             }
-            ExitCode::from(cli::ExitStatus::Failure.code() as u8)
+            ExitCode::from(cli::ExitStatus::Failure.code())
         }
     }
 }
